requirements.txt
# Core LangChain dependencies (v0.3+ with Pydantic 2)
langchain
langchain-community>=0.3.0
langchain-core>=0.3.0

# Ollama Integration
langchain-ollama>=0.1.0

# Vector Store
chromadb>=0.4.22

# Email parsing
email-reply-parser>=0.5.12

# Environment management
python-dotenv>=1.0.0

# Additional utilities
pydantic>=2.0.0

